Database Refactoring: Clean Architecture & Type Safety

+++ app/src/main/java/com/sedilant/yambol/UpdateTeamUseCaseImpl.kt
@@ -1,13 +1,13 @@
 package com.sedilant.yambol
 
-import com.sedilant.yambol.data.TeamRepository
+import com.sedilant.yambol.data.team.TeamRepository
 import com.sedilant.yambol.domain.UpdateTeamUseCase
 import javax.inject.Inject
 
 class UpdateTeamUseCaseImpl @Inject constructor(
     private val teamRepository: TeamRepository
 ) : UpdateTeamUseCase {
-    override suspend fun invoke(teamId: Int, newName: String) {
+    override suspend fun invoke(teamId: Long, newName: String) {
         teamRepository.updateTeam(teamId, newName.lowercase())
     }
 }
