+++ app/src/main/java/com/sedilant/yambol/domain/get/GetTeamObjectivesUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.sedilant.yambol.domain.get
 
-import com.sedilant.yambol.data.TeamRepository
+import com.sedilant.yambol.data.team.TeamRepository
 import com.sedilant.yambol.domain.models.TeamObjectivesDomainModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -9,7 +9,7 @@ import javax.inject.Inject
 class GetTeamObjectivesUseCaseImpl @Inject constructor(
     private val teamRepository: TeamRepository,
 ) : GetTeamObjectivesUseCase {
-    override suspend fun invoke(teamId: Int): Flow<List<TeamObjectivesDomainModel>> {
+    override suspend fun invoke(teamId: Long): Flow<List<TeamObjectivesDomainModel>> {
         return teamRepository.getTeamObjectives(teamId).map { list ->
             list.map {
                 TeamObjectivesDomainModel(
