+++ app/src/main/java/com/sedilant/yambol/domain/get/GetTrainWithTrainTaskByTrainIdUseCase.kt
@@ -3,5 +3,5 @@ package com.sedilant.yambol.domain.get
 import com.sedilant.yambol.domain.models.TrainWithTaskDomainModel
 
 interface GetTrainWithTrainTaskByTrainIdUseCase {
-    suspend operator fun invoke(trainId: Int): TrainWithTaskDomainModel
+    suspend operator fun invoke(trainId: Long): TrainWithTaskDomainModel
 }
