+++ app/src/main/java/com/sedilant/yambol/domain/insert/CreateTrainUseCase.kt
@@ -3,5 +3,5 @@ package com.sedilant.yambol.domain.insert
 import java.util.Date
 
 interface CreateTrainUseCase {
-    suspend operator fun invoke(date: Date, time: Float, concepts: List<String>, teamId: Int) : Int
+    suspend operator fun invoke(date: Date, time: Float, concepts: List<String>, teamId: Long) : Long
 }
